Break apart source a bit into modules.
