+++ src/StageVars.ts
@@ -1,0 +1,7 @@
+import { constants } from './constants';
+
+export const StageVars = {
+  initialGerms: constants.initialGerms,
+  maxInitialGermSpeed: constants.maxInitialGermSpeed,
+  spawnChance: constants.spawnChance,
+};
