+++ src/constants.ts
@@ -1,4 +1,6 @@
-export const Global = {
+export { constants, center };
+
+const constants = {
   size: 800,
   width: 800,
   height: 800,
@@ -12,3 +14,8 @@ export const Global = {
   laserSpeed: 1200,
   spawnChance: 0.25,
 } as const;
+
+const center = {
+  x: constants.width / 2,
+  y: constants.height / 2,
+}
